#111 refactor: cs Domain response package 생성
