+++ src/main/kotlin/com/psr/psr/cs/dto/response/FaqListRes.kt
@@ -1,4 +1,4 @@
-package com.psr.psr.cs.dto
+package com.psr.psr.cs.dto.response
 
 data class FaqListRes (
     val faqLists: List<FaqRes>?
