+++ src/main/kotlin/com/psr/psr/cs/dto/response/NoticeListRes.kt
@@ -1,4 +1,4 @@
-package com.psr.psr.cs.dto
+package com.psr.psr.cs.dto.response
 
 data class NoticeListRes (
     val noticeLists: List<NoticeRes>?
